Remove unused Runtime parameter from ResetEvent.set()
The Runtime parameter was never used by ResetEvent.set() (it was
immediately discarded with `_ = runtime`). Removing it simplifies
the API and fixes the Windows signal handler which was passing an
undefined Runtime pointer to work around this.

This is a breaking API change, but the signal handling feature has
not been released yet.