Fix select() to only cancel registered futures
Previously, select() would call asyncCancelWait() on all futures except
the winner, even if asyncWait() was never called on some futures. This
could happen when hitting the fast path - if a future was already
complete, we'd return early but still try to cancel all futures in the
defer block.

Now we track how many futures we've actually registered with and only
cancel those we've called asyncWait() on.