Improve Closure API to automatically provide coroutine argument
The Closure helper now automatically prepends the coroutine pointer as
the first argument when calling the wrapped function. This simplifies
the API:

Before:
  var closure = Closure(fn).init(.{ &coro, arg1, arg2 });

After:
  var closure = Closure(fn).init(.{ arg1, arg2 });

The coroutine is still passed to the function, but users don't need to
manually provide it in the init tuple. This makes the API cleaner and
less error-prone.